+++ src/Trixi.jl
@@ -235,7 +235,6 @@ export ode_norm, ode_unstable_check
 export convergence_test, jacobian_fd, jacobian_ad_forward, linear_structure
 
 export DGMulti, estimate_dt, DGMultiMesh, GaussSBP
-export VertexMappedMesh # TODO: DGMulti, v0.5. Remove deprecated VertexMappedMesh in next release
 
 export ViscousFormulationBassiRebay1, ViscousFormulationLocalDG