+++ src/solvers/dgmulti/types.jl
@@ -88,29 +88,6 @@ function DGMultiMesh(dg::DGMulti, geometric_term_type, md::MeshData{NDIMS}, boun
   return DGMultiMesh{NDIMS, geometric_term_type, typeof(md), typeof(boundary_faces)}(md, boundary_faces)
 end
 
-# TODO: DGMulti, v0.5. These constructors which use `rd::RefElemData` are now redundant and can be removed.
-function DGMultiMesh(vertex_coordinates::NTuple{NDIMS, Vector{Tv}}, EToV::Array{Ti,2}, rd::RefElemData;
-                     is_on_boundary = nothing,
-                     periodicity=ntuple(_->false, NDIMS), kwargs...) where {NDIMS, Tv, Ti}
-
-  Base.depwarn("`DGMultiMesh` constructor with `rd::RefElemData` is deprecated. Use the constructor with `dg::DGMulti` instead.",
-               :DGMultiMesh)
-  if haskey(kwargs, :is_periodic)
-    # TODO: DGMulti, v0.5. Remove deprecated keyword
-    Base.depwarn("keyword argument `is_periodic` is now `periodicity`.", :DGMultiMesh)
-    periodicity=kwargs[:is_periodic]
-  end
-
-  md = MeshData(vertex_coordinates, EToV, rd)
-  if NDIMS==1
-    md = StartUpDG.make_periodic(md, periodicity...)
-  else
-    md = StartUpDG.make_periodic(md, periodicity)
-  end
-  boundary_faces = StartUpDG.tag_boundary_faces(md, is_on_boundary)
-  return DGMultiMesh{NDIMS, typeof(rd.element_type), typeof(md), typeof(boundary_faces)}(md, boundary_faces)
-end
-
 function DGMultiMesh(triangulateIO, rd::RefElemData{2, Tri}, boundary_dict::Dict{Symbol, Int})
 
   vertex_coordinates, EToV = StartUpDG.triangulateIO_to_VXYEToV(triangulateIO)
@@ -118,9 +95,6 @@ function DGMultiMesh(triangulateIO, rd::RefElemData{2, Tri}, boundary_dict::Dict
   boundary_faces = StartUpDG.tag_boundary_faces(triangulateIO, rd, md, boundary_dict)
   return DGMultiMesh{2, typeof(rd.element_type), typeof(md), typeof(boundary_faces)}(md, boundary_faces)
 end
-
-# TODO: DGMulti, v0.5. Remove deprecated constructor
-@deprecate VertexMappedMesh(args...; kwargs...) DGMultiMesh(args...; kwargs...)
 
 # Mesh types used internally for trait dispatch
 struct Cartesian end
@@ -142,39 +116,6 @@ GeometricTermsType(mesh_type::Curved, element_type::AbstractElemShape) = NonAffi
 
 # other potential constructor types to add later: Bilinear, Isoparametric{polydeg_geo}, Rational/Exact?
 # other potential mesh types to add later: Polynomial{polydeg_geo}?
-
-"""
-  DGMultiMesh(vertex_coordinates, EToV, dg::DGMulti{NDIMS};
-              is_on_boundary=nothing,
-              periodicity=ntuple(_->false, NDIMS)) where {NDIMS, Tv}
-
-- `vertex_coordinates` is a tuple of vectors containing x,y,... components of the vertex coordinates
-- `EToV` is a 2D array containing element-to-vertex connectivities for each element
-- `dg::DGMulti` contains information associated with to the reference element (e.g., quadrature,
-  basis evaluation, differentiation, etc).
-- `is_on_boundary` specifies boundary using a `Dict{Symbol, <:Function}`
-- `periodicity` is a tuple of booleans specifying if the domain is periodic `true`/`false` in the
-  (x,y,z) direction.
-"""
-# TODO: DGMulti v0.5. Standardize order of arguments, pass in `dg` first
-function DGMultiMesh(vertex_coordinates, EToV, dg::DGMulti{NDIMS};
-                     is_on_boundary=nothing,
-                     periodicity=ntuple(_->false, NDIMS), kwargs...) where {NDIMS}
-  if haskey(kwargs, :is_periodic)
-    # TODO: DGMulti, v0.5. Remove deprecated keyword
-    Base.depwarn("keyword argument `is_periodic` is now `periodicity`.", :DGMultiMesh)
-    periodicity=kwargs[:is_periodic]
-  end
-
-  md = MeshData(vertex_coordinates, EToV, dg.basis)
-  if NDIMS == 1
-    md = StartUpDG.make_periodic(md, periodicity...)
-  else
-    md = StartUpDG.make_periodic(md, periodicity)
-  end
-  boundary_faces = StartUpDG.tag_boundary_faces(md, is_on_boundary)
-  return DGMultiMesh(dg, GeometricTermsType(VertexMapped(), dg), md, boundary_faces)
-end
 
 """
     DGMultiMesh(triangulateIO, dg::DGMulti{2, Tri}, boundary_dict::Dict{Symbol, Int})
@@ -212,12 +153,6 @@ function DGMultiMesh(dg::DGMulti{NDIMS}; cells_per_dimension,
                      coordinates_max=ntuple(_ -> one(real(dg)), NDIMS),
                      is_on_boundary=nothing,
                      periodicity=ntuple(_ -> false, NDIMS), kwargs...) where {NDIMS}
-
-  if haskey(kwargs, :is_periodic)
-    # TODO: DGMulti. Deprecate `is_periodic` in version 0.5
-    Base.depwarn("keyword argument `is_periodic` is now `periodicity`.", :DGMultiMesh)
-    periodicity=kwargs[:is_periodic]
-  end
 
   vertex_coordinates, EToV = StartUpDG.uniform_mesh(dg.basis.element_type, cells_per_dimension...)
   domain_lengths = coordinates_max .- coordinates_min